+++ checkNatspec.js
@@ -40,6 +40,8 @@ excludes = [
   "CompoundModule",
   "AaveV2Module",
   "AaveV3Borrower",
+  "lib/orbit-protocol/",
+  "lib/openzeppelin/",
 ];
 
 let anyFindings = false;