+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Upgrade to @mangrovedao/mangrove-deployments v2.1.1
 - Upgrade to @mangrovedao/context-addresses v1.2.0
+- Add Orbit lend to the routing logics
 
 # 2.1.0-0